+++ docs/src/changelog.md
@@ -17,6 +17,7 @@ If your currently installed *Quality-time* version is not the latest version, pl
 ### Fixed
 
 - Use the locale of the user's browser when exporting to PDF so that dates are formatted accordingly in the PDF. Fixes [#8381](https://github.com/ICTU/quality-time/issues/8381).
+- Show an error message when measurement entities cannot be loaded. Fixes [#10478](https://github.com/ICTU/quality-time/issues/10478).
 - When measuring average issue lead time with Jira as source, the lead times per issue would not be collected. Fixes [#10929](https://github.com/ICTU/quality-time/issues/10929).
 - Wait for all measurement entities to have been loaded before exporting to PDF. Fixes [#10992](https://github.com/ICTU/quality-time/issues/10992).